# backend/src/main/resources/static/openapi.yaml
openapi: "3.1.0"
info:
  title: "BikeTheftTracker"
  description: "BikeTheftTracker API"
  version: "1.0.0"
servers:
  - url: "http://localhost:8080/api/v1"

paths:
  /auth/register:
    post:
      tags:
        - Auth
      summary: "Register a new user"
      operationId: "registerUser"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterUserRequest"
        required: true
      responses:
        "201":
          description: "CREATED"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegisterUserResponse"
  /auth/login:
    post:
      tags:
        - Auth
      summary: "Login with an existing user"
      operationId: "login"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
  /auth/me:
    get:
      security:
        - access-token: [ ]
      tags:
        - Auth
      summary: "Get logged-in user's id, username, and roles"
      operationId: "getCurrentUser"
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDetailsResponse"
  /auth/logout:
    post:
      security:
        - access-token: [ ]
      tags:
        - Auth
      summary: "Log out user"
      operationId: "logout"
      responses:
        "204":
          description: "NO CONTENT"
          content: {}
  /auth/refresh:
    post:
      security:
        - access-token: [ ]
      tags:
        - Auth
      summary: "Get a new access token if refresh token is still valid"
      operationId: "refreshAccessToken"
      responses:
        "200":
          description: "OK"
          content: {}

  /theft-reports:
    get:
      tags:
        - TheftReports
      summary: "Returns all TheftReport map items within the visible map area."
      operationId: "getTheftReportMapItems"
      parameters:
        - name: minLon
          in: query
          required: true
          schema:
            type: number
            format: double
          description: "Minimum longitude of the map view."
        - name: minLat
          in: query
          required: true
          schema:
            type: number
            format: double
          description: "Minimum latitude of the map view."
        - name: maxLon
          in: query
          required: true
          schema:
            type: number
            format: double
          description: "Maximum longitude of the map view."
        - name: maxLat
          in: query
          required: true
          schema:
            type: number
            format: double
          description: "Maximum latitude of the map view."
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/TheftReportMapItemResponse"
    post:
      security:
        - access-token: []
      tags:
        - TheftReports
      summary: "Add a new TheftReport"
      operationId: "createTheftReport"
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: "object"
              properties:
                theftReport:
                  $ref: "#/components/schemas/CreateTheftReportRequest"
                images:
                  description: "Up to 5 images (PNG, JPEG) can be uploaded. Each file must be 5 MB or smaller"
                  type: "file"
            encoding:
              theftReport:
                contentType: "application/json"
              images:
                contentType: "image/png, image/jpeg"
      responses:
        "201":
          description: "CREATED"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TheftReportResponse"
  /theft-reports/{id}:
    get:
      tags:
        - TheftReports
      summary: "Get full details of specified TheftReport"
      operationId: "getTheftReport"
      parameters:
        - name: "id"
          in: "path"
          required: true
          description: "TheftReport id"
          schema:
            type: "string"
            format: "uuid"
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TheftReportResponse"
        "404":
          description: "NOT FOUND"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      security:
        - access-token: [ ]
      tags:
        - TheftReports
      summary: "Update logged-in user's TheftReport"
      operationId: "updateTheftReport"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "string"
            format: "uuid"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTheftReportRequest"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TheftReportResponse"

    delete:
      security:
        - access-token: [ ]
      tags:
        - TheftReports
      summary: "Delete logged-in user's TheftReport"
      operationId: "deleteTheftReport"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "string"
            format: "uuid"
      responses:
        "204":
          description: "NO CONTENT"
          content: {}
  /theft-reports/me:
    get:
      tags:
        - TheftReports
      security:
        - access-token: [ ]
      summary: "Get logged-in user's theft reports"
      operationId: "getLoggedInUsersTheftReports"
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/TheftReportResponse"
  /theft-reports/{id}/status:
    patch:
      tags:
        - TheftReports
      security:
        - access-token: [ ]
      summary: "Update status of logged-in user's TheftReport"
      operationId: "updateTheftReportStatus"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "string"
            format: "uuid"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStatusRequest"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TheftReportResponse"

  /theft-reports/{id}/sightings:
    get:
      tags:
        - Sightings
      security:
        - access-token: [ ]
      summary: "Retrieve sightings for the logged-in user's theft report"
      operationId: "getSightings"
      parameters:
        - name: "id"
          description: "TheftReport id"
          in: "path"
          required: true
          schema:
            type: "string"
            format: "uuid"
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/SightingResponse"
    post:
      tags:
        - Sightings
      security:
        - access-token: [ ]
      summary: "Create a sighting for a theft report"
      operationId: "createSighting"
      parameters:
        - name: "id"
          description: "TheftReport id"
          in: "path"
          required: true
          schema:
            type: "string"
            format: "uuid"
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: "object"
              properties:
                sighting:
                  $ref: "#/components/schemas/SightingRequest"
                image:
                  description: "One image (PNG, JPEG) can be uploaded. File must be 5 MB or smaller"
                  type: "file"
            encoding:
              theftReport:
                contentType: "application/json"
              images:
                contentType: "image/png, image/jpeg"
      responses:
        "201":
          description: "CREATED"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SightingResponse"
  /notifications:
    get:
      tags:
        - Notifications
      security:
        - access-token: [ ]
      summary: "Retrieve all unread notifications for logged-in user"
      operationId: "getUnreadNotifications"
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/NotificationResponse"
  /notifications/count:
    get:
      tags:
        - Notifications
      security:
        - access-token: [ ]
      summary: "Get unread notification count for logged-in user"
      operationId: "getUnreadNotificationsCount"
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationCountResponse"
  /notifications/{id}/read:
    patch:
      tags:
        - Notifications
      security:
        - access-token: [ ]
      summary: "Mark specified notification as read"
      operationId: "markNotificationAsRead"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "string"
            format: "uuid"
      responses:
        "200":
          description: "OK"

components:
  securitySchemes:
    accessToken:
      type: apiKey
      in: cookie
      name: access-token
    refreshToken:
      type: apiKey
      in: cookie
      name: refresh-token

  schemas:
    GeoPoint:
      type: "object"
      properties:
        longitude:
          type: "number"
          format: "double"
          example: 42.431
          minimum: -180.0
          maximum: 180
        latitude:
          type: "number"
          format: "double"
          example: 53.135
          minimum: -90.0
          maximum: 90
    TheftReportMapItemResponse:
      type: "object"
      properties:
        id:
          type: "string"
          format: "uuid"
        brand:
          type: "string"
          example: "Helkama"
        model:
          type: "string"
          example: "XS29.3"
        type:
          type: "string"
          example: "maastopyörä"
        color:
          type: "string"
          example: "sininen"
        status:
          type: "string"
          enum:
            - "ACTIVE"
            - "SIGHTED"
            - "RECOVERED"
            - "CLOSED"
        location:
          $ref: "#/components/schemas/GeoPoint"
        theftTime:
          type: "string"
          format: "date-time"
    UserResponse:
      type: "object"
      properties:
        id:
          type: "string"
          format: "uuid"
        username:
          type: "string"
          example: "test_user"
    BikeResponse:
      type: "object"
      properties:
        id:
          type: "string"
          format: "uuid"
        brand:
          type: "string"
          example: "Helkama"
        model:
          type: "string"
          example: "XS29.3"
        type:
          type: "string"
          example: "maastopyörä"
        color:
          type: "string"
          example: "sininen"
        serialNumber:
          type: "string"
          nullable: true
          example: "A12B34C56D78E"
        description:
          type: "string"
          example: "punaiset vanteet, jalka rikki"
        user:
          $ref: "#/components/schemas/UserResponse"
    TheftReportResponse:
      type: "object"
      properties:
        id:
          type: "string"
          format: "uuid"
        description:
          type: "string"
          example: "varastettu pyöräkatoksesta"
        theftTime:
          type: "string"
          format: "date-time"
        theftAddress:
          type: "string"
          nullable: true
          example: "Yliopistokatu 1, 80100 Joensuu"
        location:
          $ref: "#/components/schemas/GeoPoint"
        status:
          type: "string"
          enum:
            - "ACTIVE"
            - "SIGHTED"
            - "RECOVERED"
            - "CLOSED"
        createdAt:
          type: "string"
          format: "date-time"
        bike:
          $ref: "#/components/schemas/BikeResponse"
        images:
          type: "array"
          items:
            type: "string"
          example: [
            "http://localhost:8080/images/bikes/3fa85f64-5717-4562-b3fc-2c963f66afa6/fca2a871-53f2-4077-9c57-c6c1724c3094.jpeg",
            "http://localhost:8080/images/bikes/3fa85f64-5717-4562-b3fc-2c963f66afa6/dbf20e73-6b64-4ad8-a378-69c6cf5af17e.jpeg"
          ]
    CreateBikeRequest:
      type: "object"
      required:
        - brand
        - model
        - type
        - color
        - description
      properties:
        brand:
          type: "string"
          example: "Helkama"
          minLength: 1
          maxLength: 30
        model:
          type: "string"
          example: "XS29.3"
          minLength: 1
          maxLength: 30
        type:
          type: "string"
          example: "maastopyörä"
          minLength: 1
          maxLength: 30
        color:
          type: "string"
          example: "sininen"
          minLength: 1
          maxLength: 30
        serialNumber:
          type: "string"
          nullable: true
          example: "A12B34C56D78E"
          minLength: 1
          maxLength: 30
        description:
          type: "string"
          example: "Punaiset vanteet, jalka rikki"
          minLength: 1
          maxLength: 300
    CreateTheftReportRequest:
      type: "object"
      required:
        - description
        - theftTime
        - location
        - bike
      properties:
        description:
          type: "string"
          example: "Varastettu pyöräkatoksesta"
          minLength: 1
          maxLength: 300
        theftTime:
          type: "string"
          format: "date-time"
        theftAddress:
          type: "string"
          example: "Yliopistokatu 1, 80100 Joensuu"
          nullable: true
          maxLength: 150
        location:
          $ref: "#/components/schemas/GeoPoint"
        bike:
          $ref: "#/components/schemas/CreateBikeRequest"
    ErrorResponse:
      type: "object"
      properties:
        error:
          type: "string"
          example: "NOT FOUND"
        statusCode:
          type: "integer"
          example: 404
        message:
          type: "string"
          example: "TheftReport was not found with given id"
    RegisterUserRequest:
      type: "object"
      required:
        - username
        - password
        - email
      properties:
        username:
          type: "string"
          example: "test_user"
          minLength: 3
          maxLength: 20
          pattern: "^[a-zA-Z0-9äöåÄÖÅ](?:[a-zA-Z0-9._-äöåÄÖÅ]*[a-zA-Z0-9äöåÄÖÅ])?$"
          description: "Username must start and end with a letter or number and may contain letters, numbers, ., _, or -."
        password:
          type: "string"
          example: "VerySecurePassword123!"
          minLength: 8
          maxLength: 64
          pattern: "^(?=.*[a-z])(?=.*[A-Z])(?=.*[0-9])(?=.*[@$!%*?&^#()-_+=[]{}|:;,.<>]).*$"
          description: "Password must contain uppercase, lowercase, number and special character."
        email:
          type: "string"
          example: "test@email.com"
          minLength: 6
          maxLength: 254
          description: "Email must be valid."
    RegisterUserResponse:
      type: "object"
      properties:
        username:
          type: "string"
          example: "test_user"
        message:
          type: "string"
          example: "User created successfully. Please log in."
    LoginRequest:
      type: "object"
      required:
        - username
        - password
      properties:
        username:
          type: "string"
          example: "test_user"
          minLength: 3
          maxLength: 20
        password:
          type: "string"
          example: "VerySecurePassword123!"
          minLength: 8
          maxLength: 64
    LoginResponse:
      type: "object"
      properties:
        id:
          type: "string"
          format: "uuid"
        username:
          type: "string"
          example: "test_user"
    UserDetailsResponse:
      type: "object"
      properties:
        id:
          type: "string"
          format: "uuid"
        username:
          type: "string"
          example: "test_user"
        roles:
          type: "array"
          items:
            type: "string"
          example: ["ROLE_USER"]
    SightingResponse:
      type: "object"
      properties:
        id:
          type: "string"
          format: "uuid"
        location:
          $ref: "#/components/schemas/GeoPoint"
        description:
          type: "string"
          example: "Pyörä löytyi S-market Kauppakadun pyörätelineestä"
        reporter:
          type: "string"
          example: "test_user"
        image:
          type: "string"
          example: "http://localhost:8080/images/sightings/3fa85f64-5717-4562-b3fc-2c963f66afa6/fca2a871-53f2-4077-9c57-c6c1724c3094.jpeg"
        createdAt:
          type: "string"
          format: "date-time"
    SightingRequest:
      type: "object"
      properties:
        description:
          type: "string"
          example: "Pyörä löytyi S-market Kauppakadun pyörätelineestä"
          minLength: 1
          maxLength: 300
        location:
          $ref: "#/components/schemas/GeoPoint"
    NotificationResponse:
      type: "object"
      properties:
        id:
          type: "string"
          format: "uuid"
        type:
          type: "string"
          enum:
            - "NEW_SIGHTING"
        read:
          type: "boolean"
          example: "false"
        time:
          type: "string"
          format: "date-time"
        theftReport:
          type: "string"
          format: "uuid"
    NotificationCountResponse:
      type: "object"
      properties:
        unreadCount:
          type: "integer"
          format: "int32"
          example: 3
    UpdateBikeRequest:
      type: "object"
      properties:
        id:
          type: "string"
          format: "uuid"
        brand:
          type: "string"
          example: "Helkama"
          minLength: 1
          maxLength: 30
        model:
          type: "string"
          example: "XS29.3"
          minLength: 1
          maxLength: 30
        type:
          type: "string"
          example: "maastopyörä"
          minLength: 1
          maxLength: 30
        color:
          type: "string"
          example: "sininen"
          minLength: 1
          maxLength: 30
        serialNumber:
          type: "string"
          nullable: true
          example: "A12B34C56D78E"
          minLength: 1
          maxLength: 30
        description:
          type: "string"
          example: "punaiset vanteet, jalka rikki"
          minLength: 1
          maxLength: 300
        user:
          $ref: "#/components/schemas/UserResponse"
    UpdateTheftReportRequest:
      type: "object"
      properties:
        id:
          type: "string"
          format: "uuid"
        description:
          type: "string"
          example: "varastettu pyöräkatoksesta"
          minLength: 1
          maxLength: 300
        theftTime:
          type: "string"
          format: "date-time"
        theftAddress:
          type: "string"
          nullable: true
          example: "Yliopistokatu 1, 80100 Joensuu"
          maxLength: 150
        location:
          $ref: "#/components/schemas/GeoPoint"
        status:
          type: "string"
          enum:
            - "ACTIVE"
            - "SIGHTED"
            - "RECOVERED"
            - "CLOSED"
        createdAt:
          type: "string"
          format: "date-time"
        bike:
          $ref: "#/components/schemas/UpdateBikeRequest"
        images:
          type: "array"
          items:
            type: "string"
          example: [
            "http://localhost:8080/images/bikes/3fa85f64-5717-4562-b3fc-2c963f66afa6/fca2a871-53f2-4077-9c57-c6c1724c3094.jpeg",
            "http://localhost:8080/images/bikes/3fa85f64-5717-4562-b3fc-2c963f66afa6/dbf20e73-6b64-4ad8-a378-69c6cf5af17e.jpeg"
          ]
    UpdateStatusRequest:
      type: "object"
      properties:
        status:
          type: "string"
          enum:
            - "ACTIVE"
            - "SIGHTED"
            - "RECOVERED"
            - "CLOSED"